// Angular/project/src/app/modelos/crear-oferta-inf-admin.interface.ts
export interface infoAdminI {
  id_oferta: Number | null | undefined;
  ti_persona_captura: String | null | undefined;
  ni_persona_captura: Number | null | undefined;
  email_persona_captura: String | null | undefined;
  area_persona_captura: String | null | undefined;
  ti_persona_verifica: String | null | undefined;
  ni_persona_verifica: Number | null | undefined;
  email_persona_verifica: String | null | undefined;
  area_persona_verifica: String | null | undefined;
}